HBASE-10694 TableSkewCostFunction#cost() casts integral division result to double
git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.98@1575034 13f79535-47bb-0310-9956-ffa450edef68

+++ hbase-server/src/main/java/org/apache/hadoop/hbase/master/balancer/StochasticLoadBalancer.java
@@ -920,7 +920,7 @@ public class StochasticLoadBalancer extends BaseLoadBalancer {
     @Override
     double cost() {
       double max = cluster.numRegions;
-      double min = cluster.numRegions / cluster.numServers;
+      double min = ((double) cluster.numRegions) / cluster.numServers;
       double value = 0;
 
       for (int i = 0; i < cluster.numMaxRegionsPerTable.length; i++) {
